django-backend/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
psycopg[binary]
djangorestframework
nltk==3.9.1
